[refactor] dto 폴더 명 변경
- 기존에 변경했던 dtos -> dto가 적용되지 않은 폴더 적용
