+++ back/src/skills/dto/getGroupedSkills.dto.ts
@@ -1,6 +1,6 @@
 import { ObjectType, Field } from '@nestjs/graphql';
 
-import { CoreResponse } from '@/common/dtos/coreResponse.dto';
+import { CoreResponse } from '@/common/dto/coreResponse.dto';
 import { Skill } from '../skills.model';
 
 @ObjectType()
