+++ back/src/tags/dto/getTags.dto.ts
@@ -1,6 +1,6 @@
 import { ObjectType, Field, Int } from '@nestjs/graphql';
 
-import { CoreResponse } from '@common/dtos/coreResponse.dto';
+import { CoreResponse } from '@/common/dto/coreResponse.dto';
 import { Tag } from '../tags.model';
 
 @ObjectType()
